+++ src/sparseml/sparsification/__init__.py
@@ -20,3 +20,7 @@ algorithms to models within in ML frameworks.
 # flake8: noqa
 
 from .info import *
+from .modifier_epoch import *
+from .modifier_lr import *
+from .modifier_params import *
+from .modifier_pruning import *
